requirements.txt
alembic==1.5.7
cbor2==5.2.0 # cbor2==4.0.1
cffi==1.14.5
click==7.1.2
cryptography==3.4.6 # cryptography==2.3.1
dnspython==2.1.0
email-validator==1.1.2
Flask==1.1.2 # Flask==1.0.2
Flask-Login==0.5.0 # Flask-Login==0.4.0
Flask-Migrate==2.7.0
Flask-SQLAlchemy==2.4.4 # Flask-SQLAlchemy>=2.3.2
Flask-WTF==0.14.3 # Flask-WTF==0.14.2
future==0.18.2 # future==0.17.1
greenlet==1.0.0
gunicorn==20.0.4
idna==3.1
itsdangerous==1.1.0
Jinja2==2.11.3
Mako==1.1.4
MarkupSafe==1.1.1
pycparser==2.20
PyJWT==2.0.1
pyOpenSSL==20.0.1 # pyOpenSSL==17.5.0
python-dateutil==2.8.1
python-decouple==3.4
python-editor==1.0.4
six==1.15.0 # six==1.11.0
SQLAlchemy==1.3.23 # SQLAlchemy>=1.3.3
webauthn==0.4.7
Werkzeug==1.0.1
WTForms==2.3.3 # WTForms==2.1